refactor landing into modular components
