+++ src/components/landing/CTASection.stories.tsx
@@ -1,0 +1,5 @@
+import CTASection from "./CTASection";
+
+export default { title: "Landing/CTASection", component: CTASection };
+
+export const Default = () => <CTASection onJoinNow={() => {}} onOpenTelegram={() => {}} />;
