+++ src/components/landing/FeatureGrid.stories.tsx
@@ -1,0 +1,5 @@
+import FeatureGrid from "./FeatureGrid";
+
+export default { title: "Landing/FeatureGrid", component: FeatureGrid };
+
+export const Default = () => <FeatureGrid />;
