+++ src/components/landing/HeroSection.stories.tsx
@@ -1,0 +1,5 @@
+import HeroSection from "./HeroSection";
+
+export default { title: "Landing/HeroSection", component: HeroSection };
+
+export const Default = () => <HeroSection onOpenTelegram={() => {}} />;
