+++ src/components/landing/TestimonialsSection.stories.tsx
@@ -1,0 +1,5 @@
+import TestimonialsSection from "./TestimonialsSection";
+
+export default { title: "Landing/TestimonialsSection", component: TestimonialsSection };
+
+export const Default = () => <TestimonialsSection />;
